update translations
finish settling merge conflicts and move mockvscodeapi to mockdata folder
