+++ src/client/src/translations/whitelist_en.json
@@ -16,9 +16,6 @@
   "loginCard.signIn",
   "outputPath.outputPath",
   "selectableCard.details",
-  "selectPages.pagesTitleQuestion",
-  "selectPages.webAppTitleQuestion",
-  "summarySection.editResource",
   "about.templatesVersion",
   "about.wizardVersion",
   "azureFunctionsModal.subscriptionLabel",
@@ -85,9 +82,12 @@
   "rightSidebar.services",
   "selectBackendFramework.selectBackendFramework",
   "selectFrontendFramework.selectFrontendFramework",
+  "selectPages.pagesTitleQuestion",
+  "selectPages.webAppTitleQuestion",
   "sortablePageList.duplicateName",
   "sortablePageList.hide",
   "sortablePageList.show",
+  "summarySection.editResource",
   "azureLogin.azureTitle",
   "azureLogin.azureCardBody",
   "azureLogin.azureLoginTitle",
